requirements.txt
asgiref==3.2.10
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.7
Django==3.1.1
django-allauth==0.44.0
django-cookies-samesite==0.9.0
django-cors-headers==3.7.0
django-silk==4.1.0
django-taggit==1.4.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
gprof2dot==2021.2.21
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==20.9
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.5
